backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.10 
python-dotenv==1.0.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
